% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcardmaincontent.R
\name{dcardBoardContent}
\alias{dcardBoardContent}
\title{DCard Board Content}
\usage{
dcardBoardContent(board, posts, by_popular = F, rate_limit = 1)
}
\arguments{
\item{rate_limit}{}
}
\value{

}
\description{
Returns general post info in a specific board, where the post number and whether or not to sort by popular can be specified. The rate limit can also be altered, at the risk of being blocked by the API.
}
