% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcardcommentcontent.R
\name{dcardCommentContent}
\alias{dcardCommentContent}
\title{DCard Comment Content}
\usage{
dcardCommentContent(board, posts, by_popular = F, rate_limit = 1)
}
\arguments{
\item{rate_limit}{}
}
\value{

}
\description{
Returns general comment info and content in a specific board, where the number of posts and whether or not to sort by popular can be specified. The rate limit can also be altered, at the risk of being blocked by the API.
}
