% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcardpostcontent.R
\name{dcardPostContent}
\alias{dcardPostContent}
\title{DCard Post Content}
\usage{
dcardPostContent(board, posts, by_popular = F, rate_limit = 1)
}
\arguments{
\item{rate_limit}{}
}
\value{

}
\description{
Returns post-specific content in a board. This differs from the dcardBoardContent function, as that function only returns an excerpt of the posts, as opposed to the whole content of the post.
}
